Added new file Array+Identifiable - extension
