+++ Memorize/Grid.swift
@@ -37,16 +37,7 @@ struct Grid<Item, ItemView>: View where Item: Identifiable, ItemView: View {
     }
 }
 
-extension Array where Element: Identifiable {
-    func firstIndex(matching: Element) -> Int {
-        for index in 0..<self.count {
-            if self[index].id == matching.id {
-                return index
-            }
-        }
-        return 0 // TODO: bogus
-    }
-}
+
